+++ creme/datasets/__init__.py
@@ -1,4 +1,5 @@
+from .airline import load_airline
 from .restaurants import fetch_restaurants
 
 
-__all__ = ['fetch_restaurants']
+__all__ = ['fetch_restaurants', 'load_airline']