+++ creme/stats/rolling_mean.py
@@ -39,4 +39,4 @@ class RollingMean(rolling_sum.RollingSum):
         return 'rolling_{self.window_size}_mean'
 
     def get(self):
-        return super().get() / len(self)
+        return super().get() / len(self) if len(self) > 0 else 0