+++ docs/make_notebooks.sh
@@ -1,0 +1,1 @@
+jupyter nbconvert --execute --to html ../notebooks/*.ipynb --output-dir _build/html/notebooks