+++ CHANGELOG.md
@@ -22,7 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
     - `RollingMode`
     - `RollingSum`
     - `RollingPeakToPeak`
-- `wrap_sklearn` to `compat`
+- `creme_to_sklearn` to `compat`
 - `SVD` to `reco`
 - `BoxCoxTransformRegressor`, `TargetModifierRegressor` to `compose`
 - `iter_csv` to `stream`